requirements.txt

python-dotenv
pandas
numpy
tqdm
requests
Pillow
imagehash
boto3
chromadb==0.5.11
open-clip-torch
sentence-transformers
spotipy
pyarrow
matplotlib